+++ Sources/DTTableViewManager/DTTableViewManager.swift
@@ -29,7 +29,7 @@ import DTModelStorage
 
 /// Adopting this protocol will automatically inject `manager` property to your object, that lazily instantiates `DTTableViewManager` object.
 /// Target is not required to be `UITableViewController`, and can be a regular UIViewController with UITableView, or any other view, that contains UITableView.
-public protocol DTTableViewManageable : class
+public protocol DTTableViewManageable : AnyObject
 {
     /// Table view, that will be managed by DTTableViewManager. This property or `optionalTableView` property must be implemented in order for `DTTableViewManager` to work.
     var tableView : UITableView! { get }