+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 
+* Diffable datasources exceptions in Xcode 13 / iOS 15 with some internal restructuring.
 * Swift 5.4 / Xcode 12.5 warnings.
 
 ## [8.0.1](https://github.com/DenTelezhkin/DTTableViewManager/releases/tag/8.0.1)